+++ lib/controllers/index.js
@@ -15,6 +15,6 @@ module.exports = {
   logout: require('./logout'),
   register: require('./register'),
   verifyEmail: require('./verify-email'),
-  initiateSaml: require('./initiate-saml'),
-  verifySaml: require('./verify-saml')
+  samlRedirect: require('./saml-redirect'),
+  samlVerify: require('./saml-verify')
 };
