+++ lib/stormpath.js
@@ -181,10 +181,6 @@ module.exports.init = function (app, opts) {
     }
 
     if (web.logout.enabled) {
-      if (web.saml.enabled) {
-        addGetRoute(web.logout.uri, controllers.initiateSaml({logout: true}));
-      }
-
       addPostRoute(web.logout.uri, controllers.logout);
     }
 
@@ -216,8 +212,8 @@ module.exports.init = function (app, opts) {
     }
 
     if (web.saml.enabled) {
-      addGetRoute(web.saml.uri, controllers.initiateSaml());
-      addGetRoute(web.saml.verifyUri, controllers.verifySaml);
+      addGetRoute(web.saml.uri, controllers.samlRedirect);
+      addGetRoute(web.saml.verifyUri, controllers.samlVerify);
     }
 
     client.getApplication(config.application.href, function (err, application) {
